ship-hull-ocr-streamlit/core/config.py
import yaml
from pathlib import Path

CONFIG_PATH = Path("config.yaml")

def load_config():
    if not CONFIG_PATH.exists():
        raise FileNotFoundError(f"{CONFIG_PATH} not found!")
    with open(CONFIG_PATH, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)

CONFIG = load_config()