ship-hull-ocr-streamlit/requirements.txt
paddlepaddle==2.6.2
paddleocr==2.8.1
opencv-python>=4.9
PyYAML>=6.0
numpy>=1.26
streamlit>=1.38.0